separate bump so the command makes more semantic sense ie. `component bump` rather then `component release bump` when no release is actually performed
