+++ RELEASING.md
@@ -4,11 +4,13 @@
 
 > This method relies on the version number being incremented manually before code is pushed to Git.
 
+`npm test && component release`
+
  * Ensure all changes are made and pushed to feature branches
  * Once the feature/bug-fix is complete, rebase from master.
  * merge your changes into master
  * bump the version number manually
-   * `npm run bump` or see [api](https://github.com/skyglobal/component-helper/blob/master/API.md#bump-the-version)
+   * `component bump` or see [api](https://github.com/skyglobal/component-helper/blob/master/API.md#bump-the-version)
 
 Your `circle.yml` should look something like:
 
@@ -26,7 +28,7 @@ deployment:
   production:
     branch: master
     commands:
-      - ./node_modules/component-helper/bin/component release --version=current
+      - ./node_modules/component-helper/bin/component release current
 machine:
   node:
     version: v0.10.33
@@ -39,8 +41,8 @@ machine:
 `npm test && component release`
 
    * This will run your tests, if they fail the release will stop.
-   * Bump the version number (`patch` by default)
-     * You could specify how to bump by appending ` --version=` along with `major`, `minor`, `patch` or `prerelease`.
+   * It bumps the version number (`patch` by default)
+     * See [api](https://github.com/skyglobal/component-helper/blob/master/API.md#bump-the-version) to be specific
    * Tag the version number and push to Git
    * Push the demo `site` to github.io
    * Push the compiled assets to the S3 (if configured)
