+++ tasks/release.js
@@ -5,22 +5,6 @@ var git = require('./utils/git');
 var helper = require('./utils/config-helper');
 var component, paths, pkg;
 
-function bump(type){
-    component = helper.getConfig();
-    if (type == 'current') return Promise.resolve(component.pkg.version);
-    log.info("\nBumping version : " + type );
-    var build = require('./build');
-    var Bump = require('./utils/bump');
-    var newVersion;
-    return new Bump(['./package.json','./README.md', component.paths.source.root + '/**/version.js'], {type: type }).run()
-        .then(function(version){
-            newVersion = version;
-            return build.html({version:version});
-        }).then(function(){
-            return newVersion;
-        }).catch(log.onError);
-}
-
 function ghPagesRelease(message){
     var ghPages = require('gh-pages');
     component = helper.getConfig();
@@ -60,6 +44,7 @@ function releaseGit(version){
 }
 
 function run(type){
+    var bump = require('./bump').run;
     var bumpedVersion;
     return bump(type).then(function(version) {
         bumpedVersion = version;
@@ -73,7 +58,6 @@ function run(type){
 
 module.exports = {
     git: releaseGit,
-    bump: bump,
     'gh-pages': ghPagesRelease,
     s3: s3,
     run: run,
